+++ src/main/java/edu/illinois/library/cantaloupe/process/package-info.java
@@ -1,5 +1,0 @@
-/**
- * <p>Provides an abstraction on top of {@link java.lang.ProcessBuilder} for
- * invoking commands that read and write from/to streams.</p>
- */
-package edu.illinois.library.cantaloupe.process;