+++ src/main/java/edu/illinois/library/cantaloupe/processor/ProcessorFactory.java
@@ -21,8 +21,6 @@ public final class ProcessorFactory {
 
     private static final Set<Class<? extends Processor>> ALL_PROCESSOR_IMPLS = Set.of(
             FfmpegProcessor.class,
-            GraphicsMagickProcessor.class,
-            ImageMagickProcessor.class,
             JaiProcessor.class,
             Java2dProcessor.class,
             KakaduNativeProcessor.class,