+++ CHANGES.md
@@ -28,7 +28,8 @@
 * TurboJpegProcessor supports EXIF, IPTC, and XMP metadata.
 * FfmpegProcessor, KakaduNativeProcessor, OpenJpegProcessor, and
   PdfBoxProcessor use libjpeg-turbo to write JPEGs if it is available.
-* KakaduDemoProcessor has been removed.
+* ImageMagickProcessor, GraphicsMagickProcessor, and KakaduDemoProcessor have
+  been removed.
 
 ### Caches